refactor(qwen35): split TP executor worker runtime into tp_executor/worker.rs
Move the per-rank worker runtime cluster (TpWorker, TpWorkerState, TpWorkerPrepared,
TpStartup*, nccl watchdogs, CublasThreadGuard, bind_worker_thread) out of the 2520-line
tp_executor.rs into a flat tp_executor/worker.rs submodule (repo convention). Root
re-exports via pub(super) use worker::*; worker reaches sibling concerns via use super::*.
Item/method/field visibility set to pub(super) (used only within the tp_executor tree) so
it is warning-free. cargo check + fmt + test --lib (95/95) green.
